+++ include/client/rpc/forward_data.hpp
@@ -43,7 +43,10 @@ struct ChunkStat {
 
 // TODO once we have LEAF, remove all the error code returns and throw them as
 // an exception.
-
+std::pair<int, ssize_t>
+ecc_forward_write(const std::string& path, const void* buf, const size_t write_size,
+               const int8_t server);
+               
 std::pair<int, ssize_t>
 forward_write(const std::string& path, const void* buf, off64_t offset,
               size_t write_size, const int8_t num_copy = 0);
@@ -59,12 +62,12 @@ forward_truncate(const std::string& path, size_t current_size, size_t new_size,
 
 std::pair<int, ChunkStat>
 forward_get_chunk_stat();
-#define GKFS_USE_ECC_DISTRIBUTION 1
-#ifdef GKFS_USE_ECC_DISTRIBUTION
-std::pair <uint64_t, uint64_t > calc_op_chunks(const std::string& path, const bool append_flag,
-              const off64_t in_offset, const size_t write_size,
-              const int64_t updated_metadentry_size);
-#endif
+
+std::pair<uint64_t, uint64_t>
+calc_op_chunks(const std::string& path, const bool append_flag,
+               const off64_t in_offset, const size_t write_size,
+               const int64_t updated_metadentry_size);
+
 
 } // namespace gkfs::rpc