+++ src/client/preload.cpp
@@ -228,7 +228,7 @@ init_environment() {
 #ifdef GKFS_USE_GUIDED_DISTRIBUTION
     auto distributor = std::make_shared<gkfs::rpc::GuidedDistributor>(
             CTX->local_host_id(), CTX->hosts().size());
-#elif GKFS_USE_ECC_DISTRIBUTION
+#elif GKFS_ENABLE_EC
     auto distributor = std::make_shared<gkfs::rpc::ECCDistributor>(
             CTX->local_host_id(), CTX->hosts().size(), CTX->get_replicas());
 #else